# src/backend/src/ApplicationManagerAPI/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Application Manager API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/JobApplication": {
      "get": {
        "tags": [
          "JobApplication"
        ],
        "operationId": "JobApplication_GetAll",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "nullable": true,
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ApplicationStatus"
                }
              ]
            },
            "x-position": 1
          },
          {
            "name": "companyId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "guid",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobApplicationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "JobApplication"
        ],
        "operationId": "JobApplication_Create",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApplicationDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobApplication/{id}": {
      "get": {
        "tags": [
          "JobApplication"
        ],
        "operationId": "JobApplication_GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApplicationDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "JobApplication"
        ],
        "operationId": "JobApplication_Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobApplicationRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "JobApplication"
        ],
        "operationId": "JobApplication_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobApplication/{id}/status": {
      "patch": {
        "tags": [
          "JobApplication"
        ],
        "operationId": "JobApplication_ChangeStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobApplicationStatusRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Draft",
          "Applied",
          "Reviewing",
          "Interviewing",
          "Offer",
          "Hired",
          "Rejected"
        ],
        "enum": [
          "Draft",
          "Applied",
          "Reviewing",
          "Interviewing",
          "Offer",
          "Hired",
          "Rejected"
        ]
      },
      "CompanyDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "name": {
            "type": "string"
          },
          "industry": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "websiteUrl": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "applicationCount": {
            "type": "integer",
            "format": "int32"
          },
          "contactPersonCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "FollowUpDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "jobApplicationId": {
            "type": "string",
            "format": "guid"
          },
          "dueDateUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "completed": {
            "type": "boolean"
          }
        }
      },
...
      "InterviewDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "jobApplicationId": {
            "type": "string",
            "format": "guid"
          },
          "interviewType": {
            "$ref": "#/components/schemas/InterviewType"
          },
          "status": {
            "$ref": "#/components/schemas/InterviewStatus"
          },
          "scheduledForUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "InterviewType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Phone",
          "Video",
          "OnSite"
        ],
        "enum": [
          "Phone",
          "Video",
          "OnSite"
        ]
      },
      "InterviewStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Scheduled",
          "Completed",
          "Cancelled"
        ],
        "enum": [
          "Scheduled",
          "Completed",
          "Cancelled"
        ]
      },
...
      "JobApplicationDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "companyId": {
            "type": "string",
            "format": "guid"
          },
          "companyName": {
            "type": "string"
          },
          "positionTitle": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "appliedOn": {
            "type": "string",
            "format": "date-time"
          },
          "jobPostingUrl": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "employmentType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/EmploymentType"
              }
            ]
          },
          "workModel": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/WorkModel"
              }
            ]
          },
          "salaryMin": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "salaryMax": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "salaryCurrency": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EmploymentType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "FullTime",
          "PartTime",
          "Contract",
          "Internship"
        ],
        "enum": [
          "FullTime",
          "PartTime",
          "Contract",
          "Internship"
        ]
      },
      "WorkModel": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "OnSite",
          "Remote",
          "Hybrid"
        ],
        "enum": [
          "OnSite",
          "Remote",
          "Hybrid"
        ]
      },
      "JobApplicationDetailDto": {
        "allOf": [
          {
            "$ref": "#/components/schemas/JobApplicationDto"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "company": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CompanyDto"
                  }
                ]
              },
              "interviews": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/InterviewDto"
                }
              },
              "followUps": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FollowUpDto"
                }
              },
              "offer": {
                "nullable": true,
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/OfferDto"
                  }
                ]
              }
            }
          }
        ]
      },
      "OfferDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "jobApplicationId": {
            "type": "string",
            "format": "guid"
          },
          "status": {
            "$ref": "#/components/schemas/OfferStatus"
          },
          "annualSalary": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "OfferStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Pending",
          "Accepted",
          "Declined"
        ],
        "enum": [
          "Pending",
          "Accepted",
          "Declined"
        ]
      },
      "CreateJobApplicationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "companyId",
          "positionTitle",
          "appliedOn"
        ],
        "properties": {
          "companyId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "positionTitle": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "appliedOn": {
            "type": "string",
            "format": "date-time"
          },
          "jobPostingUrl": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "employmentType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/EmploymentType"
              }
            ]
          },
          "workModel": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/WorkModel"
              }
            ]
          },
          "salaryMin": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "salaryMax": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "salaryCurrency": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateJobApplicationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "companyId",
          "positionTitle",
          "appliedOn"
        ],
        "properties": {
          "companyId": {
            "type": "string",
            "format": "guid",
            "minLength": 1
          },
          "positionTitle": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1
          },
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "appliedOn": {
            "type": "string",
            "format": "date-time"
          },
          "jobPostingUrl": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "employmentType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/EmploymentType"
              }
            ]
          },
          "workModel": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/WorkModel"
              }
            ]
          },
          "salaryMin": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "salaryMax": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "salaryCurrency": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UpdateJobApplicationStatusRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ApplicationStatus"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}